% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coqui_condaenv.R
\name{create_piper_condaenv}
\alias{create_piper_condaenv}
\title{Install Piper Conda Environment}
\usage{
create_piper_condaenv(envname = "piper", packages = "piper-tts", ...)
}
\arguments{
\item{envname}{The name of, or path to, a \code{conda} environment.}

\item{packages}{character vector of packages to install.  ``piper-tts"`
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}.}
}
\value{
Returns the path to the Python binary associated with the
newly-created \code{conda} environment.
}
\description{
Install Piper Conda Environment
}
\examples{
\dontrun{
  if (!is_piper_installed()) {
    create_piper_condaenv()
  }
}
}
