% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_piper.R
\name{install_piper}
\alias{install_piper}
\alias{is_piper_installed}
\title{Install piper}
\usage{
install_piper(packages = NULL, pip = TRUE, ...)

is_piper_installed()
}
\arguments{
\item{packages}{character vector of packages to install. \code{"piper-tts"}
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{pip}{Boolean; use pip for package installation?
Passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:py_install]{reticulate::py_install()}}.}
}
\value{
Nothing (\code{invisible(NULL)}).  Command is called for side effects
}
\description{
Install piper
}
\examples{
\dontrun{
  if (!is_piper_installed()) {
    install_piper()
  }
}
}
