% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_download_voice.R
\name{piper_download_voice}
\alias{piper_download_voice}
\title{Download Piper Voice}
\usage{
piper_download_voice(voice, model, config, quality = NULL)
}
\arguments{
\item{voice}{name of the voice from \link{piper_voices}}

\item{model}{Path to download the model \code{.onnx} file}

\item{config}{Path to download the config \code{.json} file}

\item{quality}{quality of the voice model.  Must be specified if
multiple qualities are in \link{piper_voices}}
}
\value{
A list of the model and config paths
}
\description{
Download Piper Voice
}
\examples{
model = tempfile(fileext = ".onnx")
config = tempfile(fileext = ".json")
if (curl::has_internet()) {
  res = piper_download_voice(voice = "ljspeech", quality = "high",
                             model = model, config = config)
}
}
