% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_synthesize.R
\name{piper_synthesize}
\alias{piper_synthesize}
\title{Synthesize Text from Piper Voice}
\usage{
piper_synthesize(
  voice,
  text,
  wav_file = NULL,
  speaker_id = NULL,
  length_scale = NULL,
  noise_scale = NULL,
  noise_w = NULL,
  sentence_silence = 0
)
}
\arguments{
\item{voice}{A voice object form \link{piper_voice}}

\item{text}{Text to by synthesized}

\item{wav_file}{output file for the wav file}

\item{speaker_id}{integer of the speaker ID}

\item{length_scale}{speaking speed (default: 1.0)}

\item{noise_scale}{noise added to the generator (default: 0.667)}

\item{noise_w}{phoneme width noise (default: 0.8)}

\item{sentence_silence}{number of seconds of silence after each sentence
(default: 0.2)}
}
\value{
A path to a wave file
}
\description{
Synthesize Text from Piper Voice
}
\examples{
if (is_piper_installed()) {

}
}
