% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_voice.R
\name{piper_voice}
\alias{piper_voice}
\title{Load a Piper Voice}
\usage{
piper_voice(model, config)
}
\arguments{
\item{model}{A .onnx model file, such as en_US-lessac-medium.onnx}

\item{config}{A .onnx.json config file, such as en_US-lessac-medium.onnx.json}
}
\value{
A \code{PiperVoice} object
}
\description{
Load a Piper Voice
}
\seealso{
\link{piper_voices} and also
\url{https://github.com/sweetbbak/Neural-Amy-TTS/tree/main/models}
}
