% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_voices.R
\docType{data}
\name{piper_voices}
\alias{piper_voices}
\title{Voices for Piper}
\format{
\subsection{\code{piper_voices}}{

A data frame with 265 rows and 5 columns:
\describe{
\item{code_language}{Code for language}
\item{language}{Name of the language}
\item{voice}{Name of the voice}
\item{quality}{Country name}
\item{url_model}{URL for model file}
\item{url_config}{URL for config file}
}
}
}
\source{
\url{https://github.com/rhasspy/piper/blob/master/VOICES.md}
}
\usage{
piper_voices
}
\description{
A \code{data.frame} of the voices presented in
https://github.com/rhasspy/piper/blob/master/VOICES.md
}
\keyword{datasets}
